# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.Catalog:
    properties:
      category:
        type: string
      category_id:
        type: integer
      colors:
        items:
          type: string
        type: array
      created_at:
        type: string
      formality_level:
        type: integer
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_deleted:
        type: boolean
      material:
        type: string
      name:
        type: string
      season:
        type: string
      season_id:
        type: integer
      styles:
        items:
          type: string
        type: array
      updated_at:
        type: string
    required:
    - category_id
    - formality_level
    - gender
    - image_url
    - material
    - name
    - season_id
    type: object
  entity.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  entity.ColorType:
    properties:
      hex:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  entity.CreateCatalogRequest:
    properties:
      category_id:
        type: integer
      color_ids:
        items:
          type: integer
        type: array
      formality_level:
        type: integer
      gender:
        type: string
      image_url:
        type: string
      material:
        type: string
      name:
        type: string
      season_id:
        type: integer
      style_ids:
        items:
          type: integer
        type: array
    required:
    - category_id
    - formality_level
    - gender
    - material
    - name
    - season_id
    type: object
...
  entity.MusicType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  entity.Season:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  entity.StyleType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  entity.UpdateCatalogRequest:
    properties:
      category_id:
        minimum: 1
        type: integer
      color_ids:
        items:
          type: integer
        type: array
      formality_level:
        type: integer
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      material:
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      season_id:
        minimum: 1
        type: integer
      style_ids:
        items:
          type: integer
        type: array
    type: object
...
  swagger.ErrorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Outfit recommendation service
  title: Drip Mate API
  version: "1.0"
paths:
...
  /catalog:
    post:
      consumes:
      - multipart/form-data
      description: Creates new catalog item with image upload
      parameters:
      - description: Create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateCatalogRequest'
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create catalog item
      tags:
      - catalog
  /catalog/{id}:
    delete:
      description: Deletes item by ID
      parameters:
      - description: Item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete catalog item
      tags:
      - catalog
    get:
      description: Returns catalog item by ID
      parameters:
      - description: Item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Catalog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get catalog item
      tags:
      - catalog
    put:
      consumes:
      - multipart/form-data
      description: Updates catalog item with optional image upload
      parameters:
      - description: Item UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateCatalogRequest'
      - description: Image file
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Catalog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update catalog item
      tags:
      - catalog
  /catalog/all:
    get:
      description: Returns paginated list of catalog items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  $ref: '#/definitions/entity.Catalog'
                type: array
              total:
                type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get all catalog items
      tags:
      - catalog
...
  /reference/categories:
    get:
      description: Returns list of clothing categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      summary: Get categories
      tags:
      - reference
  /reference/colors:
    get:
      description: Returns list of available colors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ColorType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      summary: Get colors
      tags:
      - reference
  /reference/musics:
    get:
      description: Returns list of available music preferences
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.MusicType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      summary: Get music types
      tags:
      - reference
  /reference/seasons:
    get:
      description: Returns list of seasons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Season'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      summary: Get seasons
      tags:
      - reference
  /reference/styles:
    get:
      description: Returns list of available styles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.StyleType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.ErrorResponse'
      summary: Get styles
      tags:
      - reference
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"